Frankly, a lot of stuff.
I've spent close to 40 hours trying to debug the crash related to
the tutorial dialogues. A lot of messages are now exactly matching
what the client expects, but there's no improvement in the
client's behavior. It's really not obvious what causes the
tutorials to appear, but it does seem that whatever is triggering
them to appear is also causing the crashes.

The crash occurs:

1. When there are _any_ items in the inventory equipped when
   loading a lobby.
2. When opening the Main Menu when the character level is less
   than a certain threshold.
3. When loading into an area that enables the combat UI.

The last thing I suspect might be wrong is that the area where BB
stores its action bar configuration is also part of the check for
"newbie" status, because I don't actually know where that's coming
from. The number row action bar, not the command palette.

This commit also includes a proxy for connecting to servers that
use the regular Blue Burst crypto (Ephinea uses a modified algo).